# collector-api/src/main/resources/openapi/collector-openapi.yml
openapi: 3.0.3
info:
  title: API для управления коллекторской информацией
  version: 1.0.0
servers:
  - url: http://localhost:30004/api
...
paths:
...
  /debtors/{debtorId}/document:
    get:
      tags:
        - document
      summary: Получение списка документов заемщика
      operationId: getDebtorDocuments
      parameters:
        - $ref: '#/components/parameters/DebtorId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseDocumentDTO'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debtors/{debtorId}/document/file:
    post:
      tags:
        - document
      summary: Загрузка файла документа
      operationId: uploadDocumentFile
      parameters:
        - $ref: '#/components/parameters/DebtorId'
        - $ref: '#/components/parameters/DocumentType'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Файл документа
              required:
                - file
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
                example: 'Файл debtors/15/documents/national-passport.pdf загружен'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - document
      summary: Скачивание файла документа
      operationId: downloadDocumentFile
      parameters:
        - $ref: '#/components/parameters/DebtorId'
        - $ref: '#/components/parameters/DocumentType'
      responses:
        '200':
          description: File is ready for download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            Content-Type:
              schema:
                type: string
            Content-Length:
              schema:
                type: string
            Content-Disposition:
              schema:
                type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debtors/{debtorId}/notification/{agreementId}:
    get:
      tags:
        - document
      summary: Получение уведомления виде текстового файла
      description: Генерирует и возвращает текстовый файл с уведомлением для указанного заемщика и договора
      operationId: getNotification
      parameters:
        - $ref: '#/components/parameters/DebtorId'
        - $ref: '#/components/parameters/AgreementId'
      responses:
        '200':
          description: File is ready for download
          headers:
            Content-Disposition:
              schema:
                type: string
            Content-Type:
              schema:
                type: string
            Content-Length:
              schema:
                type: string
            Location:
              schema:
                type: string
          content:
            text/plain:
              schema:
                type: string
                format: binary
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    DebtorId:
      in: path
      name: debtorId
      description: 'ID заемщика'
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    AgreementId:
      in: path
      name: agreementId
      description: 'ID договора займа'
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
    DocumentType:
      in: query
      name: type
      required: true
      description: 'Тип документа заемщика для загрузки файла'
      schema:
        type: string
        example: 'national-passport'
        enum:
          - national-passport
          - international-passport
          - driver-license
          - inn
          - snils
        x-field-extra-annotation: |
          @jakarta.validation.constraints.Pattern(regexp = "registration|residential", message = "Неверный статус адреса в запросе")
  schemas:
    ErrorResponse:
      type: object
      description: 'Модель ошибки'
      properties:
        timestamp:
          type: string
          description: 'Дата и время создания ответа с ошибкой'
          format: date-time
          example: "2025-09-18T16:43:19.973103766"
        status:
          type: integer
          description: 'Код ответа'
          example: '409'
        error:
          type: string
          description: 'Расшифровка кода ответа'
          example: 'Conflict'
        message:
          type: string
          description: 'Информационное сообщение ошибки'
          example: 'В базе данных есть другой заемщик с такими документами'
        path:
          type: string
          description: 'Путь по которому был отправлен запрос'
          example: '/api/agreements'
...
    ResponseDocumentDTO:
      type: object
      description: 'Модуль данных для ответа на GET запрос, возвращаемый документ'
      properties:
        documentType:
          $ref: '#/components/schemas/DocumentType'
        documentNumber:
          $ref: '#/components/schemas/DocumentNumber'
        issueDate:
          $ref: '#/components/schemas/IssueDate'
...
    Id:
      type: integer
      format: int64
      description: 'ID число-идентификатор (64 бит)'
      example: 1
...
    DocumentType:
      type: string
      description: 'Тип документа'
      example: 'NATIONAL_PASSPORT'
      enum:
        - 'NATIONAL_PASSPORT'
        - 'INTERNATIONAL_PASSPORT'
        - 'DRIVER_LICENSE'
        - 'INN'
        - 'SNILS'
    DocumentNumber:
      type: string
      description: 'Номер документа'
      example: '4508731456'
    IssueDate:
      type: string
      description: 'Дата выдачи документа'
      format: date
      example: '2010-02-02'
...